Use resourceful routes for api trace data
